Trait rust_ofp::ofp_controller::openflow0x01::OF0x01Controller

pub trait OF0x01Controller: OfpController<Message=Message> {
    fn new() -> Self;
    fn switch_connected(&mut self,
                    sw: u64,
                    feats: SwitchFeatures,
                    stream: &mut TcpStream);
    fn switch_disconnected(&mut self, sw: u64);
    fn packet_in(&mut self,
             sw: u64,
             xid: u32,
             pkt: PacketIn,
             stream: &mut TcpStream);

    fn send_packet_out(_: u64, xid: u32, pkt: PacketOut, stream: &mut TcpStream) { ... }
    fn send_flow_mod(_: u64, xid: u32, flow: FlowMod, stream: &mut TcpStream) { ... }
    fn send_barrier_request(_: u64, xid: u32, stream: &mut TcpStream) { ... }
}

OpenFlow0x01 Controller API

OpenFlow 1.0-specific API for communicating between a controller and the dataplane.

Required Methods

fn new() -> Self

Create a new Controller.

fn switch_connected(&mut self,
sw: u64,
feats: SwitchFeatures,
stream: &mut TcpStream)

Callback invoked with sw when a switch with identifier sw connects to the controller.

fn switch_disconnected(&mut self, sw: u64)

Callback invoked with sw when a switch with identifier sw disconnects from the controller.

fn packet_in(&mut self,
sw: u64,
xid: u32,
pkt: PacketIn,
stream: &mut TcpStream)

Callback invoked when a packet pkt with transaction ID xid from switch sw arrives at the controller.

Provided Methods

fn send_packet_out(_: u64, xid: u32, pkt: PacketOut, stream: &mut TcpStream)

Send packet pkt with transaction ID xid to switch sw from the controller.

fn send_flow_mod(_: u64, xid: u32, flow: FlowMod, stream: &mut TcpStream)

Send flowmod flow with transaction ID xid to switch sw from the controller.

fn send_barrier_request(_: u64, xid: u32, stream: &mut TcpStream)

Send barrier request with transaction ID xid to switch sw from the controller. Guarantees switch sw processes messages prior to barrier before messages after.