Module rust_ofp::openflow0x01

Modules

message
Encapsulates handling of messages implementing MessageType trait.

Structs

Capabilities
Capabilities supported by the datapath.

FlowMod
Represents modifications to a flow table from the controller.

FlowRemoved
Flow removed (datapath -> controller)

Mask

PacketIn
Represents packets received by the datapath and sent to the controller.

PacketOut
Represents packets sent from the controller.

Pattern
Fields to match against flows.

PortConfig
Flags to indicate behavior of the physical port.

PortDesc
Description of a physical port.

PortFeatures
Features of physical ports available in a datapath.

PortState
Current state of a physical port. Not configurable by the controller.

PortStatus
A physical port has changed in the datapath.

SupportedActions
Actions supported by the datapath.

SwitchFeatures
Switch features.

Enums

Action
Actions associated with flows and packets.

FlowModCmd
Type of modification to perform on a flow table.

FlowRemovedReason
Reason a flow was removed from a switch

MsgCode
OpenFlow 1.0 message type codes, used by headers to identify meaning of the rest of a message.

PacketInReason
The reason a packet arrives at the controller.

Payload
The data associated with a packet received by the controller.

PortReason
What changed about a physical port.

PseudoPort
Port behavior.

StpState
STP state of a port.

Timeout
How long before a flow entry expires.

Traits

MessageType
Common API for message types implementing OpenFlow Message Codes (see MsgCode enum).