Trait rust_ofp::ofp_controller::OfpController

pub trait OfpController {
    type Message: OfpMessage;
    fn send_message(u32, Self::Message, &mut TcpStream);
    fn handle_client_connected(&mut TcpStream);
}

OpenFlow Controller

Version-agnostic API for implementing an OpenFlow controller.

Associated Types

type Message: OfpMessage

OpenFlow message type supporting the same protocol version as the controller.

Required Methods

fn send_message(u32, Self::Message, &mut TcpStream)

Send a message to the node associated with the given TcpStream.

fn handle_client_connected(&mut TcpStream)

Perform handshake and begin loop reading incoming messages from client stream.