Struct rust_ofp::ofp_header::OfpHeader

pub struct OfpHeader { /* fields omitted */ }

OpenFlow Header

The first fields of every OpenFlow message, no matter the protocol version. This is parsed to determine version and length of the remaining message, so that it can be properly handled.

Methods

impl OfpHeader

fn new(version: u8, typ: u8, length: u16, xid: u32) -> OfpHeader

Create an OfpHeader out of the arguments.

fn size() -> usize

Return the byte-size of an OfpHeader.

fn marshal(bytes: &mut Vec<u8>, header: OfpHeader)

Fills a message buffer with the header fields of an OfpHeader.

fn parse(buf: [u8; 8]) -> Self

Takes a message buffer (sized for an OfpHeader) and returns an OfpHeader.

fn version(&self) -> u8

Return the version field of a header.

fn type_code(&self) -> MsgCode

Return the OpenFlow message type code of a header.

Safety

The typ field of the self header is expected to be a u8 within the defined range of the MsgCode enum.

fn length(&self) -> usize

Return the length field of a header. Includes the length of the header itself.

fn xid(&self) -> u32

Return the xid field of a header, the transaction id associated with this packet. Replies use the same id to facilitate pairing.